Visited_nodes added to have path of every aircraft memorised
